Re-organise source code and rewrite build for Go code
- Go package code goes into ./pkg and sample code into ./example
- use fully qualified package name in examples

- Rewrite build of Go example to build locally and in a container

- Update README
