+++ data/nuc0-14393-14393.dat
