+++ data/nuc0-14393-4.4.30.dat
