+++ data/nuc0-14393-4.8.10.dat
