+++ examples/common_darwin.go
