+++ examples/common_vsock.go
