+++ examples/common_windows.go
