+++ examples/hvgoecho.go
