+++ examples/hvgostress.go
