--- go/.gitignore
@@ -1,6 +1,0 @@
-/examples/hvgoecho.darwin
-/examples/hvgoecho.linux
-/examples/hvgoecho.exe
-/examples/hvgostress.darwin
-/examples/hvgostress.linux
-/examples/hvgostress.exe
