--- go/Dockerfile
@@ -1,5 +1,0 @@
-FROM golang:alpine
-
-RUN apk update && apk add build-base
-
-WORKDIR /go/examples
