+++ pkg/vsock/vsock_linux.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package vsock
 
 import (
