+++ scripts/winsock-etw.ps1
